uploop

+++ src/ss01_java/bai_tap/USDtoVND.java
@@ -4,12 +4,12 @@ import java.util.Scanner;
 
 public class USDtoVND {
     public static void main(String[] args) {
-        int vnd = 23000;
-        int usd ;
+        double vnd = 23000;
+        double usd ;
         Scanner scanner = new Scanner(System.in);
         System.out.print("Nhập tiền USD : ");
-        usd = scanner.nextInt();
-        int quydoi = usd*vnd;
+        usd = scanner.nextDouble();
+        double quydoi = usd*vnd;
         System.out.print("Giá trị : "  + quydoi + "VND");
     }
 }
